requirements.txt
flask==2.3.3
openai==0.28.1
tqdm==4.66.1
gunicorn==21.2.0
python-dotenv==1.0.0